refactor(sessions): move hooks + utils into sessions/
Phase 3 of 4 toward issue #181. Move useSessionManager and
useRenameState (with tests) from workspace/hooks/ to sessions/hooks/.
Move pickNextVisibleSessionId and agentForSession (with tests) from
workspace/utils/ to sessions/utils/.

Update internal imports in the moved files to point at sessions/types
via '../types' (own dir). Update sessions/components and workspace
importers (WorkspaceView, TerminalZone, command-palette test) to
reach the new locations.

Refs #181
