+++ src/features/sessions/components/Card.tsx
@@ -2,7 +2,7 @@ import { type ReactElement } from 'react'
 import { Reorder } from 'framer-motion'
 import type { Session } from '../types'
 import { StatusDot } from '../../workspace/components/StatusDot'
-import { useRenameState } from '../../workspace/hooks/useRenameState'
+import { useRenameState } from '../hooks/useRenameState'
 import { formatRelativeTime } from '../../agent-status/utils/relativeTime'
 import {
   STATE_PILL_LABEL,
