+++ src/features/sessions/hooks/useSessionManager.test.ts
