+++ src/features/workspace/WorkspaceView.tsx
@@ -14,7 +14,7 @@ import { UnsavedChangesDialog } from '../editor/components/UnsavedChangesDialog'
 import { InfoBanner } from './components/InfoBanner'
 import { CommandPalette } from '../command-palette/CommandPalette'
 import { mockNavigationItems, mockSettingsItem } from './data/mockNavigation'
-import { useSessionManager } from './hooks/useSessionManager'
+import { useSessionManager } from '../sessions/hooks/useSessionManager'
 import { useResizable } from '../../hooks/useResizable'
 import { useNotifyInfo } from './hooks/useNotifyInfo'
 import { createFileSystemService } from '../files/services/fileSystemService'
