+++ src/features/workspace/components/TerminalZone.tsx
@@ -9,8 +9,8 @@ import type {
   RestoreData,
   PaneEventHandler,
   NotifyPaneReadyResult,
-} from '../hooks/useSessionManager'
-import { isOpenSessionStatus } from '../utils/pickNextVisibleSessionId'
+} from '../../sessions/hooks/useSessionManager'
+import { isOpenSessionStatus } from '../../sessions/utils/pickNextVisibleSessionId'
 
 export interface TerminalZoneProps {
   sessions: Session[]
